14.The Failure Transport
The failure_transport

Anyways, back to it! We now have a new transport called failed... which, despite its name, is the same as any other transport. If we wanted to, we could route message classes there and consume them, just like we're doing for async.

But... the purpose of this transport is different. Near the top, there's another key called failure_transport. Uncomment that and notice that this points to our new failed transport.

What does it do? Let's see it in action! First, go restart our worker:

 php bin/console messenger:consume -vv

Woh! This time, it asks us which "receiver" - which basically means which "transport" - we want to consume. A worker can read from one or many transports - something we'll talk about later with "prioritized" transports. Let's consume just the async transport - we'll handle messages from the failed transport in a different way. And actually, to make life easier, we can pass async as an argument so that it won't ask us which transport to use each time:

 php bin/console messenger:consume -vv async

Now... let's upload some images! Then... over here... pretty quickly, all 4 of those exhaust their retries and are eventually rejected from the transport. Until now, that meant that they were gone forever. But this time... that did not happen. Before removing the message from the queue, it says:

Rejected message AddPonkaToImage will be sent to the failure transport "failed"

And then... "Sending message". So, it was removed from the async transport, but it still exists because it was sent to the failed transport.

How can we see what messages have failed and try them again if we think those failure were temporary? With a couple of shiny, new console commands. Let's talk about those next.