Apparently now that we've configured a failure_transport, if handling a message still isn't working after 3 retries, instead of being sent to /dev/null, they're sent to another transport - in our case called failed. That transport is... really... the same as any other transport... and we could use the messenger:consume command to try to process those messages again.
But, there's a better way. Run:

php bin/console messenger
Seeing Messages on the Failed Queue
Hey! Shiny new commands are hiding here! Three under messenger:failed. Try out that messenger:failed:show one:

 php bin/console messenger:failed:show
Nice! There are our 4 failed messages... just sitting there wait for us to look at them. Let's pretend that we're not sure what went wrong with these messages and want to check them out. Start by passing the 115 id:

 php bin/console messenger:failed:show 115
I love this: it shows us the error message, error class and a history of the message's misadventures through our system! It failed, was redelivered to the async transport at 05, at 06 and then at 07, it finally failed and was redelivered to the failed transport.

If we add a -vv on the command...

 php bin/console messenger:failed:show 115 -vv
Now we can see a full stack trace of what happened on that exception.

This is a really powerful way to figure out what went wrong and what to do next: do we have a bug in our app that we need to fix before retrying this? Or maybe it was a temporary failure and we can try again now? Or maybe, for some reason, we just want to remove this message entirely.

If you did want to remove this without retrying, that's the messenger:failed:remove command.